9593: Moved CaseDTO files into different folder
