+++ shared/src/business/dto/cases/RestrictedCaseDTO.ts
