+++ shared/src/business/useCases/removePetitionerAndUpdateCaptionInteractor.ts
@@ -10,7 +10,7 @@ import { UnknownAuthUser } from '@shared/business/entities/authUser/AuthUser';
 import { getCaseByDocketNumber } from '@web-api/persistence/postgres/cases/getCaseByDocketNumber';
 import { updateCaseAndAssociations } from '@web-api/business/useCaseHelper/caseAssociation/updateCaseAndAssociations';
 import { withLocking } from '@web-api/persistence/postgres/utils/mutex';
-import { CaseDTO } from '../dto/docketEntries/CaseDTO';
+import { CaseDTO } from '../dto/cases/CaseDTO';
 
 /**
  * used to remove a petitioner from a case
