+++ shared/src/business/useCases/updateContactInteractor.ts
@@ -29,9 +29,9 @@ import {
   formattedOldEmailForChangeOfAddress,
 } from '@shared/business/utilities/calculateEmail';
 import { CaseFactory } from '../entities/cases/CaseFactory';
-import { CaseDTO } from '../dto/docketEntries/CaseDTO';
-import { PublicCaseDTO } from '../dto/docketEntries/PublicCaseDTO';
-import { RestrictedCaseDTO } from '../dto/docketEntries/RestrictedCaseDTO';
+import { CaseDTO } from '../dto/cases/CaseDTO';
+import { PublicCaseDTO } from '../dto/cases/PublicCaseDTO';
+import { RestrictedCaseDTO } from '../dto/cases/RestrictedCaseDTO';
 
 /**
  * updateContact
