+++ shared/src/proxies/getCaseProxy.ts
@@ -1,7 +1,7 @@
-import { CaseDTO } from '@shared/business/dto/docketEntries/CaseDTO';
+import { CaseDTO } from '@shared/business/dto/cases/CaseDTO';
 import { get } from './requests';
-import { PublicCaseDTO } from '@shared/business/dto/docketEntries/PublicCaseDTO';
-import { RestrictedCaseDTO } from '@shared/business/dto/docketEntries/RestrictedCaseDTO';
+import { PublicCaseDTO } from '@shared/business/dto/cases/PublicCaseDTO';
+import { RestrictedCaseDTO } from '@shared/business/dto/cases/RestrictedCaseDTO';
 
 /**
  * getCaseInteractor
