+++ web-api/src/business/getEntityByName.ts
@@ -11,9 +11,9 @@ import { TrialSessionWorkingCopy } from '@shared/business/entities/trialSessions
 import { User } from '@shared/business/entities/User';
 import { UserCase } from '@shared/business/entities/UserCase';
 import { UserCaseNote } from '@shared/business/entities/notes/UserCaseNote';
-import { CaseDTO } from '@shared/business/dto/docketEntries/CaseDTO';
-import { PublicCaseDTO } from '@shared/business/dto/docketEntries/PublicCaseDTO';
-import { RestrictedCaseDTO } from '@shared/business/dto/docketEntries/RestrictedCaseDTO';
+import { CaseDTO } from '@shared/business/dto/cases/CaseDTO';
+import { PublicCaseDTO } from '@shared/business/dto/cases/PublicCaseDTO';
+import { RestrictedCaseDTO } from '@shared/business/dto/cases/RestrictedCaseDTO';
 
 export const getEntityByName = (name: string) => {
   const entitiesByName = {
