+++ web-api/src/business/useCases/docketEntry/strikeDocketEntryInteractor.ts
@@ -9,7 +9,7 @@ import { getCaseByDocketNumber } from '@web-api/persistence/postgres/cases/getCa
 import { getUserById } from '@web-api/persistence/postgres/users/getUserById';
 import { applicationContext } from '@web-api/applicationContext';
 import { updateCaseAndAssociations } from '@web-api/business/useCaseHelper/caseAssociation/updateCaseAndAssociations';
-import { CaseDTO } from '@shared/business/dto/docketEntries/CaseDTO';
+import { CaseDTO } from '@shared/business/dto/cases/CaseDTO';
 
 export const strikeDocketEntryInteractor = async (
   {
