+++ web-api/src/business/useCases/docketEntry/updateDocketEntryMetaInteractor.ts
@@ -20,7 +20,7 @@ import { withLocking } from '@web-api/persistence/postgres/utils/mutex';
 import diff from 'diff-arrays-of-objects';
 import { upsertDocketEntryRelatedEntries } from '@web-api/persistence/postgres/docketEntries/upsertDocketEntryRelatedEntries';
 import { concat } from 'lodash';
-import { CaseDTO } from '@shared/business/dto/docketEntries/CaseDTO';
+import { CaseDTO } from '@shared/business/dto/cases/CaseDTO';
 
 export const updateDocketEntryMeta = async (
   applicationContext: ServerApplicationContext,
