+++ web-api/src/business/useCases/trialSessionMinutes/saveMinuteSheetToDraftsInteractor.ts
@@ -13,7 +13,7 @@ import { updateCaseAndAssociations } from '@web-api/business/useCaseHelper/caseA
 import { getCaseByDocketNumber } from '@web-api/persistence/postgres/cases/getCaseByDocketNumber';
 import { createAndUploadMinuteSheet } from '@web-api/business/useCaseHelper/trialSessionMinutes/createAndUploadMinuteSheet';
 import { getTrialSessionById } from '@web-api/persistence/postgres/trialSessions/getTrialSessionById';
-import { CaseDTO } from '@shared/business/dto/docketEntries/CaseDTO';
+import { CaseDTO } from '@shared/business/dto/cases/CaseDTO';
 
 export const saveMinuteSheetToDraftsInteractor = async (
   applicationContext: ServerApplicationContext,
