+++ web-api/src/business/useCases/trialSessions/removeCaseFromTrialInteractor.ts
@@ -16,7 +16,7 @@ import { getTrialSessionById } from '@web-api/persistence/postgres/trialSessions
 import { updateTrialSession } from '@web-api/persistence/postgres/trialSessions/updateTrialSession';
 import { removeCaseFromTrialSession } from '@web-api/persistence/postgres/trialSessions/removeCaseFromTrialSession';
 import { deleteCasesFromTrialSession } from '@web-api/persistence/postgres/trialSessions/deleteCasesFromTrialSession';
-import { CaseDTO } from '@shared/business/dto/docketEntries/CaseDTO';
+import { CaseDTO } from '@shared/business/dto/cases/CaseDTO';
 
 export const removeCaseFromTrial = async (
   applicationContext: ServerApplicationContext,
